- Refactor Custom dialog.  - Refactor AndroidManifest.xml.
