+++ app/src/main/AndroidManifest.xml
@@ -12,14 +12,14 @@
 		tools:ignore="ScopedStorage" />
 
 	<application
-		android:name=".UnlimStorageApp"
+		android:allowBackup="false"
 		android:icon="@mipmap/ic_launcher"
 		android:label="@string/app_name"
+		android:name=".UnlimStorageApp"
+		android:requestLegacyExternalStorage="true"
 		android:roundIcon="@mipmap/ic_launcher_round"
 		android:supportsRtl="true"
-		android:requestLegacyExternalStorage="true"
-		android:theme="@style/Theme.Splash"
-		android:allowBackup="false">
+		android:theme="@style/Theme.Splash">
 		<provider
 			android:name="androidx.core.content.FileProvider"
 			android:authorities="com.shov.unlimstorage.provider"
@@ -32,8 +32,8 @@
 
 		<activity
 			android:name=".MainActivity"
-			android:windowSoftInputMode="stateVisible|adjustResize"
-			android:exported="true">
+			android:exported="true"
+			android:windowSoftInputMode="stateVisible|adjustResize">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
